Carga de paciente y condiciones segun edad y riesgo

+++ src/main/java/com/JDR/Vacunassist/Repository/TurnoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.JDR.Vacunassist.Model.Turno;
 
 @Repository
-public interface TurnoRepository extends JpaRepository<Turno, Long>{
+public interface TurnoRepository extends JpaRepository<Turno, Integer>{
 
 }
